backend/requirements.txt
# Ascendra Backend — Groq Edition
# Core API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0.0
python-dotenv>=1.0.0

# LLM — FREE, no billing
groq>=0.9.0

# LinkedIn automation + anti-bot
linkedin-api>=2.1.0
playwright>=1.40.0

# Resume (ATS-Semantic Engine)
python-docx>=1.0.0
reportlab>=4.0.0

# Email
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# Utilities
requests>=2.31.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pillow>=10.0.0
apscheduler>=3.10.0
cryptography>=42.0.0
pyyaml>=6.0
rich>=13.0.0
